+++ lib/ruby_llm/agents/image/editor/dsl.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "../concerns/image_operation_dsl"
-require_relative "../generator/content_policy"
 
 module RubyLLM
   module Agents
@@ -15,7 +14,6 @@ module RubyLLM
       #   class ProductEditor < RubyLLM::Agents::ImageEditor
       #     model "gpt-image-1"
       #     size "1024x1024"
-      #     content_policy :strict
       #   end
       #
       module DSL
@@ -30,18 +28,6 @@ module RubyLLM
             @size = value
           else
             @size || inherited_or_default(:size, config.default_image_size)
-          end
-        end
-
-        # Set or get the content policy level
-        #
-        # @param level [Symbol, nil] Policy level (:none, :standard, :moderate, :strict)
-        # @return [Symbol] The content policy level
-        def content_policy(level = nil)
-          if level
-            @content_policy = level
-          else
-            @content_policy || inherited_or_default(:content_policy, :standard)
           end
         end